% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{computations}
\alias{computations}
\title{Title Computations on IndiceVals & ProtComp data}
\usage{
computations(ProtComp, IndiceVals)
}
\arguments{
\item{ProtComp}{}

\item{IndiceVals}{}
}
\value{
Returns the co-inertia analysis results
}
\description{
Title Computations on IndiceVals & ProtComp data
}
\examples{
cia <-computations(ProtComp,IndiceVals)

}
