% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndiceVals.R
\docType{data}
\name{IndiceVals}
\alias{IndiceVals}
\title{Physico-physico-chemical properties of 20 amino-acids}
\format{A data frame with 20 rows and 420 variables. Each row
contains data for a single amino-acid, and is named following
the IUPAC naming standard. Each column contains data for one
of 420 properties.}
\source{
\url{http://pbil.univ-lyon1.fr/members/lobry/repro/cabios95}
}
\usage{
IndiceVals
}
\description{
This dataset contains 420 physico-physico-chemical properties
of 20 amino-acids.
}
\keyword{datasets}
