% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtComp.R
\docType{data}
\name{ProtComp}
\alias{ProtComp}
\title{Amino-acid composition of E. coli proteins}
\format{A data frame with 20 rows and 999 variables. Each row
contains data for a single amino-acid, and is named following
the IUPAC naming standard. Each column contains data for one
E. coli protein.}
\source{
\url{http://pbil.univ-lyon1.fr/members/lobry/repro/cabios95}
}
\usage{
ProtComp
}
\description{
This dataset contains the amino-acid composition of 999 E. coli
proteins.
}
\keyword{datasets}
