% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computations.R
\name{computations}
\alias{computations}
\title{Computations on IndiceVals & ProtComp data}
\usage{
computations(ProtComp, IndiceVals, axes = 3)
}
\arguments{
\item{ProtComp}{}

\item{IndiceVals}{}

\item{axes}{The number of kept axes (defaults to 3)}
}
\value{
Returns the co-inertia analysis results
}
\description{
Computations on IndiceVals & ProtComp data
}
\examples{
cia <-computations(ProtComp, IndiceVals)

}
